+++ operaciones.h
@@ -2,12 +2,12 @@
 #ifndef OPERACIONES_H
 #define OPERACIONES_H
 
-float calc_res(float x, float y, unsigned char op);
-float suma (float,float);
-float resta (float,float);
-float prod (float,float);
-float divi (float,float);
-float expo (float,float);
+double calc_res(double x, double y, unsigned char op);
+double suma (double,double);
+double resta (double,double);
+double prod (double,double);
+double divi (double,double);
+double expo (double,double);
 
 #endif /* OPERACIONES_H */
